+++ src/core/bean/constants.ts
@@ -1,0 +1,1 @@
+export const restrictedBeanNames = ['getBean', 'getBean', 'config', 'getPrivateBean', 'getBeanConfiguration'];