+++ src/core/build-context/constants.ts
@@ -1,1 +1,3 @@
-export const PRIVATE_TOKEN = '$$$';
+import { uniqId } from '../utils/uniqId';
+
+export const PRIVATE_TOKEN = `_${uniqId()}`;