+++ src/core/transformers/getTransformerFactory.ts
@@ -2,28 +2,29 @@ import ts, { factory } from 'typescript';
 import { uniqBy } from 'lodash';
 import { isContainerAccess } from '../ts-helpers/container/isContainerAccess';
 import { replaceContainerCall } from '../ts-helpers/container/replaceContainerCall';
-import { removeDIImportsFromStatements } from '../ts-helpers/removeDIImports';
 import { removeQuotesFromString } from '../utils/removeQuotesFromString';
 import minimatch from 'minimatch';
 import { diConfig } from '../../external/config';
 import { registerAndTransformContext } from '../build-context/registerAndTransformContext';
+
+export const transformTest: Record<string, number> = {};
 
 export const getTransformerFactory = (): ts.TransformerFactory<ts.SourceFile> => context => {
     return sourceFile => {
         if (minimatch(sourceFile.fileName, diConfig.diConfigPattern!)) {
             // console.time(`Context transformed: ${sourceFile.fileName}`);
             const transformedContext = registerAndTransformContext(context, sourceFile);
+            // transformTest[sourceFile.fileName] = transformTest[sourceFile.fileName] ? transformTest[sourceFile.fileName] + 1 : 1;
+            // console.log(transformTest);
             // console.timeEnd(`Context transformed: ${sourceFile.fileName}`);
 
             return transformedContext;
         }
 
         const factoryImportsToAdd: ts.ImportDeclaration[] = [];
-        let hasContainerAccess = false;
 
         const visitor: ts.Visitor = (node => {
             if (isContainerAccess(node)) {
-                hasContainerAccess = true;
                 return replaceContainerCall(node, factoryImportsToAdd);
             }
 
@@ -31,9 +32,6 @@ export const getTransformerFactory = (): ts.TransformerFactory<ts.SourceFile> =>
         });
 
         const newSourceFile = ts.visitNode(sourceFile, visitor);
-        const filteredStatements = hasContainerAccess ?
-            removeDIImportsFromStatements(newSourceFile.statements)
-            : newSourceFile.statements;
 
         const uniqFactoryImportsToAdd = uniqBy(factoryImportsToAdd, it => {
             if (ts.isStringLiteral(it.moduleSpecifier)) {
@@ -45,7 +43,7 @@ export const getTransformerFactory = (): ts.TransformerFactory<ts.SourceFile> =>
             sourceFile,
             [
                 ...uniqFactoryImportsToAdd,
-                ...filteredStatements,
+                ...newSourceFile.statements,
             ],
             sourceFile.isDeclarationFile,
             sourceFile.referencedFiles,