+++ test/project/src/IBeans.ts
@@ -2,4 +2,8 @@ import { IUseCase } from './lib/use-case/IUseCase';
 
 export interface IBeans {
     useCase: IUseCase;
+    useCase3: IUseCase;
 }
+
+export const xz = 12232323;
+